+++ instrumentation-tests/src/androidTest/java/com/mapbox/navigation/instrumentation_tests/utils/location/MockLocationReplayer.kt
@@ -70,6 +70,13 @@ class MockLocationReplayerRule(mockLocationUpdatesRule: MockLocationUpdatesRule)
             play()
         }
     }
+
+    fun stopAndClearEvents() {
+        mapboxReplayer?.run {
+            stop()
+            clearEvents()
+        }
+    }
 }
 
 fun Location.setUpLocation(event: ReplayEventUpdateLocation) {